Breaks if no package.json is present
If there is no `package.json` available then the path is `null` and you can't stop it from breaking.

Even if you manually provide the `projectName` in options it will still break. This adds a few simple checks to prevent it from blowing up when you do provide the name.

I've also added an appropriate test, which properly fails when run without the changes to `index.js` and correctly passes once added.

Closes #6